T000 Remove empty fields in frontmatter
Previously, Gatsby would fail with this error message:

```
There was an error in your GraphQL query:

      Field "image" must not have a selection since type "String" has no
subfields.

      This can happen if you e.g. accidentally added { } to the field "image".
If you didn't expect "image" to be of type "String" make sure that your input
source and/or plugin is correct.
      However, if you expect "value" to exist, the field might be accessible in
another subfield. Please try your query in GraphiQL and use the GraphiQL
explorer to see which fields you can query and what shape they have.
```

It would call to add a schema customization, but that wouldn't really
help us in this case. Anyhow, removing the field if it's empty works.